Clean up the old multfit training code
