+++ ulmfit/datasets/utils.py
@@ -2,24 +2,17 @@
 Utility methods for data processing.
 """
 import fire
-from fastai import *
 from fastai.text import *
 
 import shutil
 import pathlib
 import tarfile
-from sklearn import model_selection
-from sacremoses import MosesTokenizer
 from typing import Dict, Tuple, List
+
+from fastai_contrib.text_data import SentencePieceTokenizer
 
 EOS = 'xxeos' # fastai does not use eos, but we do
 SEP = 'xxsep' # special separator token for NLI
-
-def replace_std_toks(x:str) -> str:
-    "Replace standard token names with fastai supported tokens"
-    # We change tokens to f'xx{token_name}' as it is not split by Moses tokenizer,
-    # while f'<{token_name}>' is being split to: '<' f'{token_name}' '>'
-    return x.replace('<unk>', UNK).replace('<bos>', BOS).replace('<eos>', EOS)
 
 PAD_TOKEN_ID = 1
 IMDB, XNLI, TRN, VAL, TST, EN = 'imdb', 'xnli', 'train', 'val', 'test', 'en'
@@ -35,124 +28,6 @@ CLASSES = ['neg', 'pos', 'unsup']
 number_match_re = re.compile(r'^([0-9]+[,.]?)+$')
 number_split_re = re.compile(r'([,.])')
 
-class MosesPreprocessingFunc():
-
-    def __init__(self, lang: str):
-        self.mt = MosesTokenizer(lang)
-
-    def __call__(self, t: str) -> str:
-        return self.mt.tokenize(t, return_str=True, escape=True)
-
-class SentencePieceTokenizer(Tokenizer):
-    "Put together rules and a tokenizer function to tokenize text with multiprocessing."
-    def __init__(self, spm_model, lang:str='en', pre_rules:ListRules=None,
-                 post_rules:ListRules=None, special_cases:Collection[str]=None, n_cpus:int=None):
-        # moses is added to preprocessing functions
-        super().__init__(self.tok_fun_with_sp, lang, pre_rules, post_rules, special_cases, n_cpus)
-        self.spm_model = spm_model
-
-    def tok_fun_with_sp(self, lang):
-        try:
-            import sentencepiece as spm
-        except ImportError:
-            raise Exception('sentencepiece module is missing: run `pip install sentencepiece`')
-        tok = BaseTokenizer(lang)
-        tok.sp = spm.SentencePieceProcessor()
-        tok.sp.Load(str(self.spm_model))
-        return tok
-
-    def process_text(self, t:str, tok:BaseTokenizer) -> List[str]:
-        "Process one text `t` with tokenizer `tok`."
-        toks = super().process_text(t, tok)
-        toks = tok.sp.EncodeAsPieces(" ".join(toks))
-        return toks
-full_char_coverage_langs = ["bg", "cs", "da", "de", "el", "en", "es", "et", "fi", "fr", "ga", "hr", "hu",
-                       "it","lt","lv","mt","nl","pl","pt","ro","sk","sl","sv"] # all European langus
-
-def get_sentencepiece(cache_dir:PathOrStr, load_text, pre_rules: ListRules=None, post_rules:ListRules=None,
-                      vocab_size:int=30000, model_type:str='unigram', input_sentence_size:int=1E7, lang='en', fixed_character_coverage=False):
-    try:
-        import sentencepiece as spm
-    except ImportError:
-        raise Exception('sentencepiece module is missing: run `pip install sentencepiece`')
-
-    cache_dir = pathlib.Path(cache_dir)
-    pre_rules = pre_rules if pre_rules is not None else defaults.text_pre_rules
-    post_rules = post_rules if post_rules is not None else defaults.text_post_rules
-
-    special_cases = defaults.text_spec_tok # + ['xxlink', 'xxuser', 'xxnumber', 'xxemoji', 'yyemoji']
-    if not os.path.isfile(cache_dir / 'spm.model') or not os.path.isfile(cache_dir / f'itos.pkl'):
-        # load the text from the train tokens file
-        text = load_text()
-        text = filter(lambda x: len(x.rstrip(" ")), text)
-        text = (reduce(lambda t, rule: rule(t), pre_rules, line) for line in text)
-        def cleanup_n_postprocess(t):
-            t = t.split()
-            for r in post_rules:
-                t = r(t)
-            return ' '.join(t)
-        text = map(cleanup_n_postprocess, text)
-        raw_text_path = cache_dir / 'all_text.txt'
-        with open(raw_text_path, 'w') as f: f.write("\n".join(text))
-
-        if fixed_character_coverage:
-            char_coverage = 0.9995
-        else:
-            char_coverage = 1 if lang in full_char_coverage_langs else 0.99
-
-        sp_params = [
-            f"--input={raw_text_path}",
-            f"--character_coverage={char_coverage}",
-            f"--unk_id={len(special_cases)}",
-            f"--pad_id=-1",
-            f"--bos_id=-1",
-            f"--eos_id=-1",
-            f"--max_sentence_length=20480",
-            f"--input_sentence_size={int(input_sentence_size)}",
-            f"--user_defined_symbols={','.join(special_cases)}",
-            f"--model_prefix={cache_dir/'spm'}",
-            f"--vocab_size={vocab_size} --model_type={model_type}"]
-        spm.SentencePieceTrainer.Train(" ".join(sp_params))
-
-        with open(cache_dir / 'spm.vocab', 'r') as f:
-            vocab = [line.split('\t')[0] for line in f.readlines()]
-
-        pickle.dump(vocab, open(cache_dir/ f'itos.pkl', 'wb'))
-    # todo add post rules
-    vocab = Vocab(pickle.load(open(cache_dir / f'itos.pkl', 'rb')))
-    # We cannot use lambdas or local methods here, since `tok_func` needs to be
-    # pickle-able in order to be called in subprocesses when multithread tokenizing
-    tokenizer = SentencePieceTokenizer(cache_dir/'spm.model',
-                                lang=lang,
-                                pre_rules=pre_rules,
-                                post_rules=post_rules)
-    return {'tokenizer': tokenizer, 'vocab': vocab}
-
-def get_sentencepiece_fastai(cache_dir: PathOrStr,  pre_rules: ListRules = None,
-                          post_rules: ListRules = None,
-                          vocab_size: int = 30000, lang='en'):
-    cache_dir = pathlib.Path(cache_dir)
-
-    sp_model = cache_dir / 'spm.model'
-    if not sp_model.is_file():
-        sp_model = None
-
-    sp_vocab = cache_dir / 'spm.vocab'
-    if not sp_vocab.is_file():
-        sp_vocab = None
-
-    processor = SPProcessor(
-        pre_rules=pre_rules,
-        post_rules=post_rules,
-        mark_fields=True,
-        vocab_sz=vocab_size,
-        sp_model=sp_model,
-        sp_vocab=sp_vocab,
-        lang=lang,
-        tmp_dir=cache_dir.absolute()  # absolute make sure that dataset path is not added as prefix
-    )
-    return {'processor': processor}
-
 def clear_cache_directory(path:PathOrStr, cache_name:str='tmp'):
     path = pathlib.Path(path)
     shutil.rmtree(path / cache_name)
@@ -165,7 +40,6 @@ def get_texts(path):
             labels.append(idx)
     return np.array(texts), np.array(labels)
 
-
 def ensure_paths_exists(*paths, message="One or more required files cannot be found."):
     error = False
     for path in paths:
@@ -174,12 +48,6 @@ def ensure_paths_exists(*paths, message="One or more required files cannot be fo
             error = True
     if error:
         raise FileNotFoundError(message)
-
-def get_data_folder() -> Path:
-    """
-    return data folder to use for future processing
-    """
-    return (pathlib.Path(__file__).parent.parent / "data")
 
 def get_scripts_folder():
     """
@@ -339,8 +207,6 @@ def read_file(file_path, outname=None):
         df.to_csv(file_path.parent / f'{outname}.csv', header=False, index=False)
     return df
 
-
-
 def read_whitespace_file(filepath):
     """Reads a file and prepares the tokens."""
     tokens = []
@@ -356,7 +222,6 @@ class DataStump:
         self.ids = ids
         self.loss_func = F.cross_entropy
 
-
 def validate(model, ids, bptt=2000):
     """
     Return the validation loss and perplexity of a model
@@ -369,7 +234,7 @@ def validate(model, ids, bptt=2000):
     model.eval()
     model.reset()
     total_loss, num_examples = 0., 0
-    for inputs, targets in tqdm(data):
+    for inputs, targets in data:
         outputs, raws, outs = model(to_device(inputs, None))
         p_vocab = F.softmax(outputs, 1)
         for i, pv in enumerate(p_vocab):
@@ -378,7 +243,6 @@ def validate(model, ids, bptt=2000):
         num_examples += len(inputs)
     mean = total_loss / num_examples  # divide by total number of tokens
     return mean, np.exp(mean)
-
 
 class TextReader():
     """ Returns a language model iterator that iterates through batches that are of length N(bptt,5)
